Adapt `solana deploy`
